% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_df.R
\name{double_df}
\alias{double_df}
\title{Check your lr information base on double data frame or tissue}
\usage{
double_df(db, ligand, deg_lig, deg_rec, nor_lig, nor_rec, sample_info)
}
\arguments{
\item{db}{A lr database according your requirement by \code{data (db, packages = "lrcheck")}.}

\item{ligand}{Ligand name base on you research.}

\item{deg_lig}{Ligand-derived difference expr matrix/data frame,col1 is log2FC,
col2 is padj and col3 is label about gene change such as,up,down or stable.}

\item{deg_rec}{Receptor-derived difference expr matrix/data frame,col1 is log2FC,
col2 is padj and col3 is label about gene change such as,up,down or stable.}

\item{nor_lig}{Ligand-derived gene expression matrix or data frame by normalized.}

\item{nor_rec}{Receptor-derived gene expression matrix or data frame by normalized.}

\item{sample_info}{Sample information table,col1 is sample name,
col2 is group name.}
}
\value{
A list object
}
\description{
Return a list include lr database,difference,expression levels and boxplot
when you import ligand,deg df,expression df and sample information.
}
\examples{
double_df (db = mouse_db, ligand = "Btc",
deg_lig = deg_ligand, deg_rec = deg_receptor,
nor_lig = nor_ligand, nor_rec = nor_receptor,
sample_info = sample_info)
}
