% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_expr_plot.R
\name{lr_expr_plot}
\alias{lr_expr_plot}
\title{Boxplot for LR expression}
\usage{
lr_expr_plot(d, e, f)
}
\arguments{
\item{d}{sample information,col1 is sample name,col2 is group name, table format}

\item{e}{ligand}

\item{f}{expression df}
}
\value{
a boxplot
}
\description{
Boxplot for LR expression
}
\examples{
lr_expr_plot (d = sample_info, e = ligand, f = nor_dat)
}
