% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_df.R
\name{single_df}
\alias{single_df}
\title{Check your lr information base on single data frame or tissue}
\usage{
single_df(db, ligand, deg, nor, sample_info)
}
\arguments{
\item{db}{A lr database according your requirement by \code{data (db, packages = "lrcheck")}.}

\item{ligand}{Ligand name base on you research}

\item{deg}{Difference expr matrix/data frame,col1 is log2FC,
col2 is padj and col3 is label about gene change such as,up,down or stable.}

\item{nor}{Gene expression matrix or data frame by normalized.}

\item{sample_info}{Sample information table,col1 is sample name,
col2 is group name.}
}
\value{
A list object
}
\description{
Return a list include lr database,difference,expression levels and boxplot
when you import ligand,deg df,expression df and sample information.
}
\examples{
single_df(db = mouse_db,
ligand = "Mstn",
deg = deg_dat,
nor = nor_dat,
sample_info = sample_info)
}
